+++ azext_iot/operations/hub.py
@@ -32,7 +32,6 @@ from azext_iot.common.shared import (
 )
 from azext_iot.iothub.providers.discovery import IotHubDiscovery
 from azext_iot.common.utility import (
-    shell_safe_json_parse,
     read_file_content,
     validate_key_value_pairs,
     url_encode_dict,
@@ -260,11 +259,18 @@ def _assemble_auth(auth_method, pk, sk):
     )
 
     auth = None
-    if auth_method in [DeviceAuthType.shared_private_key.name, DeviceAuthApiType.sas.value]:
+    if auth_method in [
+        DeviceAuthType.shared_private_key.name,
+        DeviceAuthApiType.sas.value,
+    ]:
         auth = AuthenticationMechanism(
-            symmetric_key=SymmetricKey(primary_key=pk, secondary_key=sk), type=DeviceAuthApiType.sas.value
+            symmetric_key=SymmetricKey(primary_key=pk, secondary_key=sk),
+            type=DeviceAuthApiType.sas.value,
         )
-    elif auth_method in [DeviceAuthType.x509_thumbprint.name, DeviceAuthApiType.selfSigned.value]:
+    elif auth_method in [
+        DeviceAuthType.x509_thumbprint.name,
+        DeviceAuthApiType.selfSigned.value,
+    ]:
         if not pk:
             raise ValueError("primary thumbprint required with selfSigned auth")
         auth = AuthenticationMechanism(
@@ -273,8 +279,13 @@ def _assemble_auth(auth_method, pk, sk):
             ),
             type=DeviceAuthApiType.selfSigned.value,
         )
-    elif auth_method in [DeviceAuthType.x509_ca.name, DeviceAuthApiType.certificateAuthority.value]:
-        auth = AuthenticationMechanism(type=DeviceAuthApiType.certificateAuthority.value)
+    elif auth_method in [
+        DeviceAuthType.x509_ca.name,
+        DeviceAuthApiType.certificateAuthority.value,
+    ]:
+        auth = AuthenticationMechanism(
+            type=DeviceAuthApiType.certificateAuthority.value
+        )
     else:
         raise ValueError("Authorization method {} invalid.".format(auth_method))
     return auth
@@ -868,7 +879,11 @@ def _handle_module_update_params(parameters):
 
 
 def _parse_auth(parameters):
-    valid_auth = [DeviceAuthApiType.sas.value, DeviceAuthApiType.selfSigned.value, DeviceAuthApiType.certificateAuthority.value]
+    valid_auth = [
+        DeviceAuthApiType.sas.value,
+        DeviceAuthApiType.selfSigned.value,
+        DeviceAuthApiType.certificateAuthority.value,
+    ]
     auth = parameters["authentication"].get("type")
     if auth not in valid_auth:
         raise CLIError("authentication.type must be one of {}".format(valid_auth))
@@ -1380,31 +1395,67 @@ def _process_config_content(content, config_type):
 
 def _validate_payload_schema(content):
     import json
-    from azext_iot.constants import EDGE_DEPLOYMENT_SCHEMA_2_PATH as schema_path
+    from os.path import join
     from azext_iot.models.validators import JsonSchemaType, JsonSchemaValidator
+    from azext_iot.constants import EDGE_DEPLOYMENT_ROOT_SCHEMAS_PATH as root_schema_path
+    from azext_iot.common.utility import shell_safe_json_parse
 
-    if not exists(schema_path):
-        logger.info("Invalid schema path %s, skipping validation...", schema_path)
-        return
+    EDGE_AGENT_SCHEMA_PATH = "azure-iot-edgeagent-deployment-{}.json"
+    EDGE_HUB_SCHEMA_PATH = "azure-iot-edgehub-deployment-{}.json"
+    EDGE_SCHEMA_PATH_DICT = {
+        "$edgeAgent": EDGE_AGENT_SCHEMA_PATH,
+        "$edgeHub": EDGE_HUB_SCHEMA_PATH,
+    }
 
-    logger.info("Validating deployment payload...")
-    schema_content = str(read_file_content(schema_path))
+    modules_content = content["modulesContent"]
+    system_modules_for_validation = ["$edgeAgent", "$edgeHub"]
 
-    try:
-        schema_content = shell_safe_json_parse(schema_content)
-    except CLIError:
-        logger.info("Issue parsing Edge deployment schema, skipping validation...")
-        return
+    for sys_module in system_modules_for_validation:
+        if sys_module in modules_content:
+            if (
+                "properties.desired" in modules_content[sys_module]
+                and "schemaVersion"
+                in modules_content[sys_module]["properties.desired"]
+            ):
+                target_schema_ver = modules_content[sys_module][
+                    "properties.desired"
+                ]["schemaVersion"]
+                target_schema_def_path = join(root_schema_path, f"{EDGE_SCHEMA_PATH_DICT[sys_module].format(target_schema_ver)}")
 
-    v = JsonSchemaValidator(schema_content, JsonSchemaType.draft4)
-    errors = v.validate(content)
-    if errors:
-        # Pretty printing schema validation errors
-        raise CLIError(
-            json.dumps({"validationErrors": errors}, separators=(",", ":"), indent=2)
-        )
+                logger.info("Attempting to fetch schema content from %s...", target_schema_def_path)
+                if not exists(target_schema_def_path):
+                    logger.info("Invalid schema path %s, skipping validation...", target_schema_def_path)
+                    continue
 
-    return
+                try:
+                    target_schema_def = str(read_file_content(target_schema_def_path))
+                    target_schema_def = shell_safe_json_parse(target_schema_def)
+                except Exception:
+                    logger.info(
+                        "Unable to fetch schema content from %s skipping validation...",
+                        target_schema_def_path,
+                    )
+                    continue
+
+                logger.info(f"Validating {sys_module} of deployment payload against schema...")
+                to_validate_content = {
+                    sys_module: modules_content[sys_module]
+                }
+                draft_version = JsonSchemaType.draft4
+                if "$schema" in target_schema_def and "/draft-07/" in target_schema_def["$schema"]:
+                    draft_version = JsonSchemaType.draft7
+
+                v = JsonSchemaValidator(target_schema_def, draft_version)
+                errors = v.validate(to_validate_content)
+                if errors:
+                    # Pretty printing schema validation errors
+                    raise CLIError(
+                        json.dumps(
+                            {"validationErrors": errors},
+                            separators=(",", ":"),
+                            indent=2,
+                        )
+                    )
 
 
 def iot_hub_configuration_update(
@@ -2029,7 +2080,10 @@ def _build_device_or_module_connection_string(entity, key_type="primary"):
             if key_type == "primary"
             else auth["symmetricKey"]["secondaryKey"]
         )
-    elif auth_type in [DeviceAuthApiType.certificateAuthority.value.lower(), DeviceAuthApiType.selfSigned.value.lower()]:
+    elif auth_type in [
+        DeviceAuthApiType.certificateAuthority.value.lower(),
+        DeviceAuthApiType.selfSigned.value.lower(),
+    ]:
         key = "x509=true"
     else:
         raise CLIError("Unable to form target connection string")
@@ -2130,8 +2184,14 @@ def _iot_device_send_message(
     import os
 
     device = _iot_device_show(target, device_id)
-    if device and device.get("authentication", {}).get("type", "") != DeviceAuthApiType.sas.value:
-        raise CLIError('D2C send message command only supports symmetric key auth (SAS) based devices')
+    if (
+        device
+        and device.get("authentication", {}).get("type", "")
+        != DeviceAuthApiType.sas.value
+    ):
+        raise CLIError(
+            "D2C send message command only supports symmetric key auth (SAS) based devices"
+        )
 
     msgs = []
     if properties:
@@ -2521,8 +2581,14 @@ def iot_simulate_device(
         if protocol_type == ProtocolType.mqtt.name:
 
             device = _iot_device_show(target, device_id)
-            if device and device.get("authentication", {}).get("type", "") != DeviceAuthApiType.sas.value:
-                raise CLIError('MQTT simulation is only supported for symmetric key auth (SAS) based devices')
+            if (
+                device
+                and device.get("authentication", {}).get("type", "")
+                != DeviceAuthApiType.sas.value
+            ):
+                raise CLIError(
+                    "MQTT simulation is only supported for symmetric key auth (SAS) based devices"
+                )
 
             wrap = mqtt_client_wrap(
                 target=target,
@@ -2614,6 +2680,7 @@ def iot_device_export(
     resource_group_name=None,
 ):
     from azext_iot._factory import iot_hub_service_factory
+
     client = iot_hub_service_factory(cmd.cli_ctx)
     discovery = IotHubDiscovery(cmd)
     target = discovery.get_target(
@@ -2639,21 +2706,31 @@ def iot_device_export(
             authentication_type=storage_authentication_type,
         )
 
-        user_identity = identity not in [None, '[system]']
-        if user_identity and storage_authentication_type != AuthenticationType.identityBased.name:
+        user_identity = identity not in [None, "[system]"]
+        if (
+            user_identity
+            and storage_authentication_type != AuthenticationType.identityBased.name
+        ):
             raise CLIError(
                 "Device export with user-assigned identities requires identity-based authentication [--storage-auth-type]"
             )
         # Track 2 CLI SDKs provide support for user-assigned identity objects
-        if ensure_iothub_sdk_min_version(IOTHUB_TRACK_2_SDK_MIN_VERSION) and user_identity:
-            from azure.mgmt.iothub.models import ManagedIdentity  # pylint: disable=no-name-in-module
+        if (
+            ensure_iothub_sdk_min_version(IOTHUB_TRACK_2_SDK_MIN_VERSION)
+            and user_identity
+        ):
+            from azure.mgmt.iothub.models import (
+                ManagedIdentity,
+            )  # pylint: disable=no-name-in-module
+
             export_request.identity = ManagedIdentity(user_assigned_identity=identity)
 
         # if the user supplied a user-assigned identity, let them know they need a new CLI/SDK
         elif user_identity:
             raise CLIError(
-                "Device export with user-assigned identities requires a dependency of azure-mgmt-iothub>={}"
-                .format(IOTHUB_TRACK_2_SDK_MIN_VERSION)
+                "Device export with user-assigned identities requires a dependency of azure-mgmt-iothub>={}".format(
+                    IOTHUB_TRACK_2_SDK_MIN_VERSION
+                )
             )
 
         return client.export_devices(
@@ -2715,20 +2792,30 @@ def iot_device_import(
             authentication_type=storage_authentication_type,
         )
 
-        user_identity = identity not in [None, '[system]']
-        if user_identity and storage_authentication_type != AuthenticationType.identityBased.name:
+        user_identity = identity not in [None, "[system]"]
+        if (
+            user_identity
+            and storage_authentication_type != AuthenticationType.identityBased.name
+        ):
             raise CLIError(
                 "Device import with user-assigned identities requires identity-based authentication [--storage-auth-type]"
             )
         # Track 2 CLI SDKs provide support for user-assigned identity objects
-        if ensure_iothub_sdk_min_version(IOTHUB_TRACK_2_SDK_MIN_VERSION) and user_identity:
-            from azure.mgmt.iothub.models import ManagedIdentity  # pylint: disable=no-name-in-module
+        if (
+            ensure_iothub_sdk_min_version(IOTHUB_TRACK_2_SDK_MIN_VERSION)
+            and user_identity
+        ):
+            from azure.mgmt.iothub.models import (
+                ManagedIdentity,
+            )  # pylint: disable=no-name-in-module
+
             import_request.identity = ManagedIdentity(user_assigned_identity=identity)
         # if the user supplied a user-assigned identity, let them know they need a new CLI/SDK
         elif user_identity:
             raise CLIError(
-                "Device import with user-assigned identities requires a dependency of azure-mgmt-iothub>={}"
-                .format(IOTHUB_TRACK_2_SDK_MIN_VERSION)
+                "Device import with user-assigned identities requires a dependency of azure-mgmt-iothub>={}".format(
+                    IOTHUB_TRACK_2_SDK_MIN_VERSION
+                )
             )
 
         return client.import_devices(